@@ -17,7 +17,7 @@
     ],
     "require": {
         "php": ">=7.2.5",
-        "egulias/email-validator": "^2.1.10|^3",
+        "egulias/email-validator": "^2.1.10|^3|^4",
         "psr/event-dispatcher": "^1",
         "psr/log": "^1|^2|^3",
         "symfony/deprecation-contracts": "^2.1|^3",
@@ -27,7 +27,7 @@
         "symfony/service-contracts": "^1.1|^2|^3"
     },
     "require-dev": {
-        "symfony/http-client-contracts": "^1.1|^2|^3",
+        "symfony/http-client": "^4.4|^5.0|^6.0",
         "symfony/messenger": "^4.4|^5.0|^6.0"
     },
     "conflict": {
