@@ -23,18 +23,20 @@
         "symfony/polyfill-php80": "^1.16"
     },
     "require-dev": {
-        "egulias/email-validator": "^2.1.10|^3.1",
+        "egulias/email-validator": "^2.1.10|^3.1|^4",
         "phpdocumentor/reflection-docblock": "^3.0|^4.0|^5.0",
         "symfony/dependency-injection": "^4.4|^5.0|^6.0",
+        "symfony/process": "^5.4|^6.4",
         "symfony/property-access": "^4.4|^5.1|^6.0",
         "symfony/property-info": "^4.4|^5.1|^6.0",
-        "symfony/serializer": "^5.2|^6.0"
+        "symfony/serializer": "^5.4.35|~6.3.12|^6.4.3"
     },
     "conflict": {
         "egulias/email-validator": "~3.0.0",
         "phpdocumentor/reflection-docblock": "<3.2.2",
         "phpdocumentor/type-resolver": "<1.4.0",
-        "symfony/mailer": "<4.4"
+        "symfony/mailer": "<4.4",
+        "symfony/serializer": "<5.4.35|>=6,<6.3.12|>=6.4,<6.4.3"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Mime\\": "" },
