@@ -310,6 +310,7 @@ class SitemapPlugin extends Plugin
                 $include_lang = $this->multilang_skiplang_prefix !== $lang;
                 $location = $page->canonical($include_lang);
                 $url = $page->url(false, $include_lang);
+                $lastmod = !empty($header->sitemap['lastmod']) ? strtotime($header->sitemap['lastmod']) : $page->modified();
 
                 $lang_route = [
                     'title' => $page->title(),
@@ -318,7 +319,7 @@ class SitemapPlugin extends Plugin
                     'lang' => $lang,
                     'translated' => in_array($lang, $page_languages),
                     'location' => $location,
-                    'lastmod' => date($this->datetime_format, $page->modified()),
+                    'lastmod' => date($this->datetime_format, $lastmod),
                     'longdate' => date('Y-m-d\TH:i:sP', $page->date()),
                     'shortdate' => date('Y-m-d', $page->date()),
                     'timestamp' => $page->date(),
